python_requirements.txt
paho.mqtt
requests
flask